Adicionada funcionalidade de taxa como solicitado 	- A thread que controlava a taxa por cliente foi removida 	- Agora existe um arquivo que guarda o IP de cada cliente que se conecta ao lado de um valor referente à sua taxa de bytes/seg 	- Caso seja a primeira vez que o cliente se conecta com aquele IP, é setada uma taxa randômica para ele e salva no arquivo 	- o arquivo 'gerencia.c' é o responsável por fazer a manipulação do arquivo 'gerenciaTaxas' e verificar a taxa do cliente 	- Foi mantido o arquivo 'config' que guarda a taxa maxima do servidor e é gerado pelo comando './Server -a <taxa>' 	- Caso um cliente tente logar e sua taxa de bytes/seg irá ser superior à taxa que o servidor comporta, ele não poderá logar e receberá uma mensagem de erro 	- A comparação da taxa de bytes para permitir entrada é feita com base na taxa disponível, que é a (taxa_maxima - taxa_usada)

+++ arquivo.c
@@ -2,26 +2,26 @@
 #include <stdlib.h>
 #include <string.h>
 void arquivo(char maxTaxaS[]);
-float lerArquivo();
+int lerArquivo();
 
 void arquivo(char maxTaxaS[]){
-    float maxTaxa;
+    int maxTaxa;
     FILE *arq;
-    maxTaxa = atof(maxTaxaS);
-    printf("Taxa máxima recebida: %.2f B/s.\n", maxTaxa);
+    maxTaxa = atoi(maxTaxaS);
+    printf("Taxa máxima recebida: %i B/s.\n", maxTaxa);
     arq = fopen("config", "w");
-    fprintf(arq, "%.2f", maxTaxa);
+    fprintf(arq, "%i", maxTaxa);
     fclose(arq);
     printf("Arquivo 'config' criado com sucesso.\n");
     system(EXIT_SUCCESS);
 }
 
-float lerArquivo(){
-    float maxTaxa;
+int lerArquivo(){
+    int maxTaxa;
     FILE *arq;
     arq = fopen("config", "r");
-    fscanf(arq, "%f", &maxTaxa);
-    printf("Taxa de transferencia máxima setada para: %.2f Bytes por segundo.\n", maxTaxa);
+    fscanf(arq, "%i", &maxTaxa);
+    printf("Taxa de transferencia máxima setada para: %i Bytes por segundo.\n", maxTaxa);
     fclose(arq);
     return maxTaxa;
 }
